moved slack integration urls to a different file
